requirements.txt
potassium
requests
openai-whisper
